Moved assertState into the shared support file.
