+++ tests/cypress/support/autocomplete-test-support.ts
@@ -2,6 +2,7 @@
 // and also retrieve the shared CSS and HTML elements IDs exposed
 // by Strype via the Window object of the app.
 import {WINDOW_STRYPE_HTMLIDS_PROPNAME, WINDOW_STRYPE_SCSSVARS_PROPNAME} from "@/helpers/sharedIdCssWithTests";
+import {cleanFromHTML} from "../support/test-support";
 
 export let scssVars: {[varName: string]: string};
 export let strypeElIds: {[varName: string]: (...args: any[]) => string};
@@ -153,5 +154,41 @@ export function checkAutocompleteSorted(acIDSel: string, isInFuncCallFrame: bool
             // Replace opening bracket onwards as we want to check it's sorted by function name, ignoring params:
             .then((items) => [...items].map((item) => item.innerText.toLowerCase().replace(new RegExp("\\(.*"), "")))
             .should("beLocaleSorted");
+    });
+}
+
+// Checks that the first labelslot in the given frame has content equivalent to expectedState (with a dollar indicating cursor position),
+// and equivalent to expectedStateWithPlaceholders if you count placeholders as the text for blank spans
+// If the last parameter is missing, it's assumed that expectedStateWithPlaceholders is the same as expectedState
+// (but without the dollar)
+export function assertState(frameId: number, expectedState : string, expectedStateWithPlaceholders?: string) : void {
+    expectedStateWithPlaceholders = expectedStateWithPlaceholders ?? expectedState.replaceAll("$", "");
+    withSelection((info) => {
+        cy.get("#" + strypeElIds.getFrameHeaderUID(frameId) + " #" + strypeElIds.getFrameLabelSlotsStructureUID(frameId, 0) + " ." + scssVars.labelSlotInputClassName).then((parts) => {
+            let content = "";
+            let contentWithPlaceholders = "";
+            for (let i = 0; i < parts.length; i++) {
+                const p : any = parts[i];
+                let text = cleanFromHTML(p.value || p.textContent || "");
+
+                // If the text for a span is blank, use the placeholder since that's what the user will be seeing:
+                if (!text) {
+                    // Get rid of zero-width spaces (trim() doesn't seem to do this):
+                    contentWithPlaceholders += p.getAttribute("placeholder")?.replace(/\u200B/g,"") ?? "";
+                }
+                else {
+                    contentWithPlaceholders += text;
+                }
+
+                // If we're the focused slot, put a dollar sign in to indicate the current cursor position:
+                if (info.id === p.getAttribute("id") && info.cursorPos >= 0) {
+                    text = text.substring(0, info.cursorPos) + "$" + text.substring(info.cursorPos);
+                }
+
+                content += text;
+            }
+            expect(content).to.equal(expectedState);
+            expect(contentWithPlaceholders).to.equal(expectedStateWithPlaceholders);
+        });
     });
 }
